TICKET : list, add = OK ; edit = WIP, try to download files
