+++ src/AppBundle/Resources/public/sass/_form.scss
@@ -1,6 +1,5 @@
 .form-all {
-  margin-top : 4rem;
-  padding-top: 4rem;
+  margin: 4rem 0 0 0;
   color: $paragraph-color;
   min-height: 100%;
   font-family: Arial, sans-serif;
