+++ src/AppBundle/Resources/views/Ticket/addTicket.html.twig
@@ -1,18 +1,12 @@
 {% extends ('@App/backoffice-base.html.twig') %}
 
 {% block body %}
-
     <section class="form-all">
-
         <div class="row">
             <div class="col-xs-12 col-lg-6 col-lg-offset-3">
                 {{ form(form) }}
-
                 {#{{ dump(test) }}#}
             </div>
         </div>
     </section>
-
-
-
 {% endblock %}
