+++ web/assets/css/backoffice-main.css
@@ -3,8 +3,7 @@
 /*page background color*/
 /*background color for boxes*/
 .form-all {
-  margin-top: 4rem;
-  padding-top: 4rem;
+  margin: 4rem 0 0 0;
   color: #545454;
   min-height: 100%;
   font-family: Arial, sans-serif; }
